Bits and bobs left out

+++ Musketeers2.0/Assets/Scripts/Unit.cs
@@ -7,9 +7,9 @@ public class Unit : MonoBehaviour
     
     public string unitName;
     public int unitLevel;
-    public int damage;
+    public float damage;
     public int maxHP;
-    public int currentHP;
+    public float currentHP;
     public int dangerLevel;
     public List<string> abilites = new List<string>{""};
     
@@ -17,7 +17,7 @@ public class Unit : MonoBehaviour
 
 
 
-    public bool IsDead(int currentHP)
+     public bool IsDead(float currentHP)
     {
         if (currentHP <= 0)
         {
@@ -29,13 +29,13 @@ public class Unit : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         currentHP -= dmg;
         
     }
 
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         currentHP += amount;
         if (currentHP > maxHP)
